Rename index.js to core.js
