+++ core.d.ts
