+++ package.json
@@ -16,8 +16,8 @@
 		"test": "xo && ava && tsd"
 	},
 	"files": [
-		"index.js",
-		"index.d.ts",
+		"core.js",
+		"core.d.ts",
 		"supported.js",
 		"util.js"
 	],
